chore(components): encapsulate data fetching logic into a custom hook
