+++ utils/index.js
@@ -1,2 +1,3 @@
 export { default as buildCounterChartConfig } from './buildCounterChartConfig'
 export { default as buildTimeSeriesChartConfig } from './buildTimeSeriesChartConfig'
+export { default as useDataFetching } from './useDataFetching'
